requirements.txt
numpy
Pillow
torch>=1.4
pandas
tabulate
tqdm